Rearrange commands
