+++ components/common/CharmEditor/components/inlinePalette/editorItems/text.tsx
@@ -2,24 +2,58 @@
 import type { EditorState, Transaction } from '@bangle.dev/pm';
 import { Fragment, setBlockType } from '@bangle.dev/pm';
 import { rafCommandExec } from '@bangle.dev/utils';
+import ArrowForwardIcon from '@mui/icons-material/ArrowForwardIos';
 import ChatBubbleIcon from '@mui/icons-material/ChatBubble';
 import ChatOutlinedIcon from '@mui/icons-material/ChatOutlined';
-import CodeIcon from '@mui/icons-material/Code';
+import DescriptionOutlinedIcon from '@mui/icons-material/DescriptionOutlined';
+import FormatListBulletedIcon from '@mui/icons-material/FormatListBulleted';
+import FormatListNumberedIcon from '@mui/icons-material/FormatListNumbered';
+import HorizontalRuleIcon from '@mui/icons-material/HorizontalRule';
+import LibraryAddCheckIcon from '@mui/icons-material/LibraryAddCheck';
+import DatabaseIcon from '@mui/icons-material/TableChart';
 import TextFieldsIcon from '@mui/icons-material/TextFields';
+import type { PageType, SpaceOperation } from '@prisma/client';
 import { TextSelection } from 'prosemirror-state';
+import type { PluginKey } from 'prosemirror-state';
+
+import type { SpacePermissionFlags } from 'lib/permissions/spaces';
 
 import { insertNode, isAtBeginningOfLine } from '../../../utils';
 import * as bulletList from '../../bulletList';
+import type { NestedPagePluginState } from '../../nestedPage';
+import { nestedPageSuggestMarkName } from '../../nestedPage';
 import * as orderedList from '../../orderedList';
 import paragraph from '../../paragraph';
-import {
-  isList
-} from '../commands';
+import { isList } from '../commands';
 import { palettePluginKey } from '../config';
 import { replaceSuggestionMarkWith } from '../inlinePalette';
-import type { PaletteItemTypeNoGroup } from '../paletteItem';
+import type { PaletteItemTypeNoGroup, PromisedCommand } from '../paletteItem';
+
+interface ItemsProps {
+  addNestedPage: () => Promise<void>;
+  disableNestedPage: boolean;
+  nestedPagePluginKey?: PluginKey<NestedPagePluginState>;
+  userSpacePermissions?: SpacePermissionFlags;
+  pageType?: PageType;
+}
 
 const iconSize = 30;
+
+function createTableCell (state: EditorState, text: string) {
+  return state.schema.nodes.table_cell.create(undefined, Fragment.fromArray([
+    state.schema.nodes.paragraph.create(undefined, Fragment.fromArray([
+      state.schema.text(text)
+    ]))
+  ]));
+}
+
+function createTableHeader (state: EditorState, text: string) {
+  return state.schema.nodes.table_cell.create({ header: true }, Fragment.fromArray([
+    state.schema.nodes.paragraph.create(undefined, Fragment.fromArray([
+      state.schema.text(text)
+    ]))
+  ]));
+}
 
 const { convertToParagraph } = paragraph;
 const {
@@ -35,8 +69,10 @@ const setHeadingBlockType = (level: number) => (state: EditorState, dispatch: ((
   return setBlockType(type, { level })(state, dispatch);
 };
 
-export function items (): PaletteItemTypeNoGroup[] {
-  return [
+export function items (props: ItemsProps): PaletteItemTypeNoGroup[] {
+  const { addNestedPage, disableNestedPage, nestedPagePluginKey, pageType, userSpacePermissions } = props;
+
+  const paletteItems: PaletteItemTypeNoGroup[] = [
     {
       uid: 'paraConvert',
       keywords: ['paragraph', 'text'],
@@ -69,27 +105,101 @@ export function items (): PaletteItemTypeNoGroup[] {
         };
       }
     },
-    {
-      uid: 'code',
-      title: 'Code',
-      icon: <CodeIcon sx={{
+    ...(pageType !== 'card_template' && !disableNestedPage ? [{
+      uid: 'insert-page',
+      title: 'Insert page',
+      requiredSpacePermission: 'createPage' as SpaceOperation,
+      keywords: ['page'],
+      icon: <DescriptionOutlinedIcon sx={{
         fontSize: iconSize
       }}
       />,
-      description: 'Insert a code block in the line below',
+      description: 'Insert a new page',
+      editorExecuteCommand: (() => {
+        return (async (state, dispatch, view) => {
+          await addNestedPage();
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+        }) as PromisedCommand;
+      })
+    }] : []),
+    {
+      uid: 'todoListConvert',
+      title: 'Todo List',
+      icon: <LibraryAddCheckIcon sx={{
+        fontSize: iconSize
+      }}
+      />,
+      keywords: ['todo', 'lists', 'checkbox', 'checked'],
+      description: 'Convert the current block to todo list',
       editorExecuteCommand: () => {
         return (state, dispatch, view) => {
-          if (view) {
-            rafCommandExec(view, (_state, _dispatch) => {
-              if (isAtBeginningOfLine(_state)) {
-                return setBlockType(_state.schema.nodes.codeBlock, { language: 'Javascript' })(_state, _dispatch);
-              }
-              return insertNode(_state, _dispatch, state.schema.nodes.codeBlock.create(
-                { language: 'Javascript' },
-                Fragment.fromArray([])
-              ));
-            });
-          }
+          rafCommandExec(view!, (_state, _dispatch, _view) => {
+            setBlockType(_state.schema.nodes.paragraph)(_state, _dispatch);
+            return toggleTodoList()(_view!.state, _view!.dispatch, _view);
+          });
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+        };
+      }
+    },
+    ...Array.from({ length: 3 }, (_, i) => {
+      const level = i + 1;
+      return {
+        uid: `headingConvert${level}`,
+        icon: <svg stroke='currentColor' fill='currentColor' strokeWidth={0} viewBox='0 0 512 512' height={iconSize} width={iconSize} xmlns='http://www.w3.org/2000/svg'><path d='M448 96v320h32a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16H320a16 16 0 0 1-16-16v-32a16 16 0 0 1 16-16h32V288H160v128h32a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16H32a16 16 0 0 1-16-16v-32a16 16 0 0 1 16-16h32V96H32a16 16 0 0 1-16-16V48a16 16 0 0 1 16-16h160a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16h-32v128h192V96h-32a16 16 0 0 1-16-16V48a16 16 0 0 1 16-16h160a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16z' /></svg>,
+        title: `Heading ${level}`,
+        description: `Convert the current block to heading level ${level}`,
+        disabled: (state) => {
+          const result = isList()(state);
+          return result;
+        },
+        editorExecuteCommand: () => {
+          return (state, dispatch, view) => {
+            rafCommandExec(view!, setHeadingBlockType(level));
+            return replaceSuggestionMarkWith(palettePluginKey, '')(
+              state,
+              dispatch,
+              view
+            );
+          };
+        }
+      } as PaletteItemTypeNoGroup;
+    }),
+    {
+      uid: 'insertSimpleTable',
+      icon: <DatabaseIcon sx={{
+        fontSize: iconSize
+      }}
+      />,
+      title: 'Table',
+      keywords: ['table'],
+      description: 'Insert a simple table below',
+      editorExecuteCommand: () => {
+        return (state, dispatch, view) => {
+          rafCommandExec(view!, (_state, _dispatch) => {
+            return insertNode(_state, _dispatch, _state.schema.nodes.table.create(
+              undefined,
+              Fragment.fromArray([
+                _state.schema.nodes.table_row.create(undefined, Fragment.fromArray([
+                  createTableHeader(_state, 'Header 1'),
+                  createTableHeader(_state, 'Header 2'),
+                  createTableHeader(_state, 'Header 3')
+                ])),
+                _state.schema.nodes.table_row.create(undefined, Fragment.fromArray([
+                  createTableCell(_state, 'Cell 1'),
+                  createTableCell(_state, 'Cell 2'),
+                  createTableCell(_state, 'Cell 3')
+                ]))
+              ])
+            ));
+          });
           return replaceSuggestionMarkWith(palettePluginKey, '')(
             state,
             dispatch,
@@ -99,36 +209,84 @@ export function items (): PaletteItemTypeNoGroup[] {
       }
     },
     {
-      uid: 'callout',
-      title: 'Callout',
-      icon: <ChatBubbleIcon sx={{
+      uid: 'bulletListConvert',
+      title: 'Bullet List',
+      icon: <FormatListBulletedIcon sx={{
         fontSize: iconSize
       }}
       />,
-      description: 'Insert a callout block in the line below',
+      keywords: ['unordered', 'lists'],
+      description: 'Convert the current block to bullet list',
       editorExecuteCommand: () => {
         return (state, dispatch, view) => {
+          rafCommandExec(view!, (_state, _dispatch, _view) => {
+            setBlockType(_state.schema.nodes.paragraph)(_state, _dispatch);
+            return toggleBulletList()(_view!.state, _view!.dispatch, _view);
+          });
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+        };
+      }
+    },
+    {
+      uid: 'orderedListConvert',
+      icon: <FormatListNumberedIcon sx={{
+        fontSize: iconSize
+      }}
+      />,
+      title: 'Ordered List',
+      keywords: ['numbered', 'lists'],
+      description: 'Convert the current block to ordered list',
+      editorExecuteCommand: () => {
+        return (state, dispatch, view) => {
+          rafCommandExec(view!, (_state, _dispatch, _view) => {
+            setBlockType(_state.schema.nodes.paragraph)(_state, _dispatch);
+            return toggleOrderedList()(_view!.state, _view!.dispatch, _view);
+          });
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+        };
+      }
+    },
+    {
+      uid: 'insertDisclosure',
+      icon: <ArrowForwardIcon sx={{ fontSize: iconSize }} />,
+      title: 'Toggle List/Heading',
+      keywords: ['summary', 'disclosure', 'toggle', 'collapse'],
+      description: 'Insert a summary and content',
+      editorExecuteCommand: () => {
+        return (state, dispatch, view) => {
+
           rafCommandExec(view!, (_state, _dispatch) => {
 
-            const node = _state.schema.nodes.blockquote.create(
-              undefined,
-              Fragment.fromArray([
-                _state.schema.nodes.paragraph.create(undefined, Fragment.fromArray([]))
-              ])
-            );
+            const { $from, $to } = _state.selection;
+            const range = $from.blockRange($to);
 
-            if (_dispatch && isAtBeginningOfLine(_state)) {
+            if (_dispatch && range) {
               const tr = _state.tr;
-              tr.replaceSelectionWith(node);
-              // move cursor to block
-              const offset = tr.selection.$head.end(1); // param 1 is node deep
-              const resolvedPos = tr.doc.resolve(offset);
+              tr.wrap(range, [
+                {
+                  type: _state.schema.nodes.disclosureDetails
+                }
+              ]);
+              tr.insert(range.start + 1, _state.schema.nodes.disclosureSummary.createChecked(null, Fragment.fromArray([
+                _state.schema.nodes.paragraph.create(undefined, Fragment.fromArray([]))
+              ])));
+              const resolvedPos = tr.doc.resolve(range.start + 1);
+
               tr.setSelection(TextSelection.near(resolvedPos));
+
               _dispatch(tr);
-              return true;
             }
-            return insertNode(_state, _dispatch, node);
+            return true;
           });
+
           return replaceSuggestionMarkWith(palettePluginKey, '')(
             state,
             dispatch,
@@ -176,28 +334,103 @@ export function items (): PaletteItemTypeNoGroup[] {
         };
       }
     },
-    ...Array.from({ length: 3 }, (_, i) => {
-      const level = i + 1;
-      return {
-        uid: `headingConvert${level}`,
-        icon: <svg stroke='currentColor' fill='currentColor' strokeWidth={0} viewBox='0 0 512 512' height={iconSize} width={iconSize} xmlns='http://www.w3.org/2000/svg'><path d='M448 96v320h32a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16H320a16 16 0 0 1-16-16v-32a16 16 0 0 1 16-16h32V288H160v128h32a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16H32a16 16 0 0 1-16-16v-32a16 16 0 0 1 16-16h32V96H32a16 16 0 0 1-16-16V48a16 16 0 0 1 16-16h160a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16h-32v128h192V96h-32a16 16 0 0 1-16-16V48a16 16 0 0 1 16-16h160a16 16 0 0 1 16 16v32a16 16 0 0 1-16 16z' /></svg>,
-        title: `Heading ${level}`,
-        description: `Convert the current block to heading level ${level}`,
-        disabled: (state) => {
-          const result = isList()(state);
-          return result;
-        },
-        editorExecuteCommand: () => {
-          return (state, dispatch, view) => {
-            rafCommandExec(view!, setHeadingBlockType(level));
-            return replaceSuggestionMarkWith(palettePluginKey, '')(
-              state,
-              dispatch,
-              view
+    {
+      uid: 'divider',
+      title: 'Divider',
+      keywords: ['divider', 'hr'],
+      icon: <HorizontalRuleIcon sx={{ fontSize: iconSize }} />,
+      description: 'Display horizontal rule',
+      editorExecuteCommand: () => {
+        return (state, dispatch, view) => {
+          // Execute the animation
+          rafCommandExec(view!, (_state, _dispatch) => {
+            const node = _state.schema.nodes.horizontalRule.create({ track: [] });
+            if (_dispatch && isAtBeginningOfLine(state)) {
+              _dispatch(_state.tr.replaceSelectionWith(node));
+              return true;
+            }
+            return insertNode(_state, _dispatch, node);
+          });
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+
+        };
+      }
+    },
+    {
+      uid: 'link-to-page',
+      title: 'Link to page',
+      keywords: ['link', 'page'],
+      icon: <DescriptionOutlinedIcon sx={{
+        fontSize: iconSize
+      }}
+      />,
+      description: 'Link to a new page',
+      editorExecuteCommand: (() => {
+        return (async (state, dispatch, view) => {
+          if (nestedPagePluginKey) {
+            const nestedPagePluginState = nestedPagePluginKey.getState(state);
+            if (nestedPagePluginState) {
+              replaceSuggestionMarkWith(palettePluginKey, state.schema.text(' ', state.schema.marks[nestedPageSuggestMarkName].create({})), true)(
+                state,
+                dispatch,
+                view
+              );
+            }
+            return false;
+          }
+          return false;
+        }) as PromisedCommand;
+      })
+    },
+    {
+      uid: 'callout',
+      title: 'Callout',
+      icon: <ChatBubbleIcon sx={{
+        fontSize: iconSize
+      }}
+      />,
+      description: 'Insert a callout block in the line below',
+      editorExecuteCommand: () => {
+        return (state, dispatch, view) => {
+          rafCommandExec(view!, (_state, _dispatch) => {
+
+            const node = _state.schema.nodes.blockquote.create(
+              undefined,
+              Fragment.fromArray([
+                _state.schema.nodes.paragraph.create(undefined, Fragment.fromArray([]))
+              ])
             );
-          };
-        }
-      } as PaletteItemTypeNoGroup;
-    })
+
+            if (_dispatch && isAtBeginningOfLine(_state)) {
+              const tr = _state.tr;
+              tr.replaceSelectionWith(node);
+              // move cursor to block
+              const offset = tr.selection.$head.end(1); // param 1 is node deep
+              const resolvedPos = tr.doc.resolve(offset);
+              tr.setSelection(TextSelection.near(resolvedPos));
+              _dispatch(tr);
+              return true;
+            }
+            return insertNode(_state, _dispatch, node);
+          });
+          return replaceSuggestionMarkWith(palettePluginKey, '')(
+            state,
+            dispatch,
+            view
+          );
+        };
+      }
+    }
   ];
+
+  const allowedDynamicOtherItems = paletteItems.filter(paletteItem => {
+    return !paletteItem.requiredSpacePermission
+      || (paletteItem.requiredSpacePermission && userSpacePermissions?.[paletteItem.requiredSpacePermission]);
+  });
+
+  return allowedDynamicOtherItems;
 }
